requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
faiss-cpu==1.13.0
filelock==3.19.1
fsspec==2025.10.0
hf-xet==1.3.2
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
